Add trips combination

+++ fivethreads-trip/src/main/java/lt/fivethreads/services/TripServiceImplementation.java
@@ -1,13 +1,8 @@
 package lt.fivethreads.services;
 
-import lt.fivethreads.entities.Trip;
-import lt.fivethreads.entities.TripAcceptance;
-import lt.fivethreads.entities.TripMember;
-import lt.fivethreads.entities.TripStatus;
-import lt.fivethreads.entities.request.CreateTripForm;
-import lt.fivethreads.entities.request.EditTripInformation;
-import lt.fivethreads.entities.request.TripDTO;
-import lt.fivethreads.entities.request.TripMemberDTO;
+import lt.fivethreads.entities.*;
+import lt.fivethreads.entities.request.*;
+import lt.fivethreads.exception.AccessRightProblem;
 import lt.fivethreads.exception.TripIsNotEditable;
 import lt.fivethreads.exception.WrongTripData;
 import lt.fivethreads.mapper.TripMapper;
@@ -44,10 +39,13 @@ public class TripServiceImplementation implements TripService {
     @Autowired
     TripFilesService tripFilesService;
 
-    public TripDTO createTrip(CreateTripForm form) throws WrongTripData {
+    @Autowired
+    UserService userService;
+
+    public TripDTO createTrip(CreateTripForm form, String organizer_email) throws WrongTripData {
         tripValidation.checkFnishStartDates(form.getStartDate(), form.getFinishDate(), "Finish date is earlier than start date.");
         tripValidation.checkStartDateToday(form.getStartDate());
-        Trip trip = tripMapper.ConvertCreateTripFormToTrip(form);
+        Trip trip = tripMapper.ConvertCreateTripFormToTrip(form, organizer_email);
         for (TripMember tripMember : trip.getTripMembers()) {
             tripValidation.validateTripMember(tripMember);
             tripMember.setTripAcceptance(TripAcceptance.PENDING);
@@ -88,10 +86,14 @@ public class TripServiceImplementation implements TripService {
         return tripDTO;
     }
 
-    public TripMemberDTO addNewTripMember(TripMemberDTO tripMemberDTO, Long tripID) {
+    public TripMemberDTO addNewTripMember(TripMemberDTO tripMemberDTO, Long tripID, String organizer_email) throws AccessRightProblem, TripIsNotEditable {
         TripMember tripMember = tripMemberMapper.convertTripMemberDTOtoTripMember(tripMemberDTO);
-        tripMember.setTripAcceptance(TripAcceptance.PENDING);
         Trip trip = tripRepository.findByID(tripID);
+        if (!trip.getOrganizer().getEmail().equals(organizer_email)) {
+            throw new AccessRightProblem("Only organizer can edit or delete the trip");
+        }
+        tripMember.setTripAcceptance(TripAcceptance.PENDING);
+
         tripMember.setTrip(trip);
         tripValidation.validateTripMember(tripMember);
         if (trip.getTripStatus() == TripStatus.COMPLETED) {
@@ -102,29 +104,37 @@ public class TripServiceImplementation implements TripService {
         return tripMemberMapper.convertTripMemberToTripMemberDTO(tripMember);
     }
 
-    public void deleteTrip(Long tripID) {
+    public void deleteTrip(Long tripID, String organizer_email) throws AccessRightProblem, TripIsNotEditable {
         if (tripFilesService.checkIfDocumentsExist(tripID)) {
             throw new TripIsNotEditable("Trip cannot be deleted because financial documents exist.");
         }
         Trip trip = tripRepository.findByID(tripID);
-        for (TripMember tripMember: trip.getTripMembers()
-             ) {
+        if (!trip.getOrganizer().getEmail().equals(organizer_email)) {
+            throw new AccessRightProblem("Only organizer can edit or delete the trip");
+        }
+        for (TripMember tripMember : trip.getTripMembers()
+        ) {
             createNotificationService.createNotificationDeleted(tripMember, "Trip was deleted.");
         }
         tripRepository.deleteTrip(trip);
     }
 
 
-    public TripDTO editTripInformation(EditTripInformation editTripInformation) {
+    public TripDTO editTripInformation(EditTripInformation editTripInformation, String organizer_email) throws AccessRightProblem, TripIsNotEditable {
         if (tripFilesService.checkIfDocumentsExist(editTripInformation.getId())) {
             throw new TripIsNotEditable("Trip cannot be deleted because financial documents exist.");
         }
         Trip trip = tripRepository.findByID(editTripInformation.getId());
+        if (!trip.getOrganizer().getEmail().equals(organizer_email)) {
+            throw new AccessRightProblem("Only organizer can edit or delete the trip");
+        }
         if (!trip.getStartDate().equals(editTripInformation.getStartDate())
                 || !trip.getFinishDate().equals(editTripInformation.getFinishDate())
-                || !trip.getArrival().equals(editTripInformation.getArrival())
-                || !trip.getDeparture().equals(editTripInformation.getDeparture())) {
-            for (TripMember tripMember: trip.getTripMembers()
+                || trip.getArrival().getLongitude() != editTripInformation.getArrival().getLongitude()
+                || trip.getArrival().getLatitude() != editTripInformation.getArrival().getLatitude()
+                || trip.getDeparture().getLatitude() != editTripInformation.getDeparture().getLatitude()
+                || trip.getDeparture().getLatitude() != editTripInformation.getDeparture().getLatitude()) {
+            for (TripMember tripMember : trip.getTripMembers()
             ) {
                 createNotificationService.createNotificationForApprovalTripMember(tripMember, "Trip's information was changed. Trip is waiting for your approval.");
             }
@@ -132,8 +142,32 @@ public class TripServiceImplementation implements TripService {
         trip.setTripStatus(TripStatus.NOTSTARTED);
         trip.setStartDate(editTripInformation.getStartDate());
         trip.setFinishDate(editTripInformation.getFinishDate());
-        trip.setArrival(editTripInformation.getArrival());
-        trip.setDeparture(editTripInformation.getDeparture());
+        Address arrival = new Address();
+        arrival.setCity(editTripInformation.getArrival().getCity());
+        arrival.setCountry(editTripInformation.getArrival().getCountry());
+        arrival.setHouseNumber(editTripInformation.getArrival().getHouseNumber());
+        arrival.setFlatNumber(editTripInformation.getArrival().getFlatNumber());
+        arrival.setLatitude(editTripInformation.getArrival().getLatitude());
+        arrival.setLongitude(editTripInformation.getArrival().getLongitude());
+        arrival.setStreet(editTripInformation.getArrival().getStreet());
+        Address departure = new Address();
+        departure.setCity(editTripInformation.getDeparture().getCity());
+        departure.setCountry(editTripInformation.getDeparture().getCountry());
+        departure.setHouseNumber(editTripInformation.getDeparture().getHouseNumber());
+        departure.setFlatNumber(editTripInformation.getDeparture().getFlatNumber());
+        departure.setLatitude(editTripInformation.getDeparture().getLatitude());
+        departure.setLongitude(editTripInformation.getDeparture().getLongitude());
+        departure.setStreet(editTripInformation.getDeparture().getStreet());
+        trip.setArrival(arrival);
+        trip.setDeparture(departure);
+        tripRepository.updateTrip(trip);
+        return tripMapper.converTripToTripDTO(trip);
+    }
+
+    public TripDTO changeOrganizer(ChangeOrganizer changeOrganizer) {
+        Trip trip = tripRepository.findByID(changeOrganizer.getId());
+        User organizer = userService.getUserByEmail(changeOrganizer.getOrganizer_email());
+        trip.setOrganizer(organizer);
         tripRepository.updateTrip(trip);
         return tripMapper.converTripToTripDTO(trip);
     }
